Create a separate layer for all Moveable instances
