+++ assets/src/edit-story/components/canvas/page.js
@@ -12,9 +12,10 @@ import { useCallback, useEffect, useState } from '@wordpress/element';
  * Internal dependencies
  */
 import { useStory } from '../../app';
+import SelectionMovable from '../selection';
+import MovableLayer from '../movable/movableLayer';
 import useCanvas from './useCanvas';
 import Element from './element';
-import Movable from './../movable';
 
 const Background = styled.div.attrs( { className: 'container' } )`
 	background-color: ${ ( { theme } ) => theme.colors.fg.v1 };
@@ -63,29 +64,31 @@ function Page() {
 
 	return (
 		<Background>
-			{ currentPage && currentPage.elements.map( ( { id, ...rest } ) => {
-				const isSelected = Boolean( selectedElement && selectedElement.id === id );
+			<MovableLayer>
+				{ currentPage && currentPage.elements.map( ( { id, ...rest } ) => {
+					const isSelected = Boolean( selectedElement && selectedElement.id === id );
 
-				return (
-					<Element
-						key={ id }
-						setNodeForElement={ setNodeForElement }
-						isEditing={ editingElement === id }
-						element={ { id, ...rest } }
-						isSelected={ isSelected }
-						handleSelectElement={ handleSelectElement }
-						forwardedRef={ isSelected ? setTargetEl : null }
+					return (
+						<Element
+							key={ id }
+							setNodeForElement={ setNodeForElement }
+							isEditing={ editingElement === id }
+							element={ { id, ...rest } }
+							isSelected={ isSelected }
+							handleSelectElement={ handleSelectElement }
+							forwardedRef={ isSelected ? setTargetEl : null }
+						/>
+					);
+				} ) }
+
+				{ selectedElement && targetEl && (
+					<SelectionMovable
+						selectedElement={ selectedElement }
+						targetEl={ targetEl }
+						pushEvent={ pushEvent }
 					/>
-				);
-			} ) }
-
-			{ selectedElement && targetEl && (
-				<Movable
-					selectedElement={ selectedElement }
-					targetEl={ targetEl }
-					pushEvent={ pushEvent }
-				/>
-			) }
+				) }
+			</MovableLayer>
 		</Background>
 	);
 }
